Moved replicated log predicates into their own file
