+++ tests/js/server/replication2/replication2-supervision-replicated-state-target-cluster.js
@@ -29,6 +29,7 @@ const db = arangodb.db;
 const lh = require("@arangodb/testutils/replicated-logs-helper");
 const sh = require("@arangodb/testutils/replicated-state-helper");
 const spreds = require("@arangodb/testutils/replicated-state-predicates");
+const lpreds = require("@arangodb/testutils/replicated-log-predicates");
 
 const database = "replication2_supervision_test_db";
 
@@ -178,7 +179,7 @@ const replicatedStateSuite = function () {
 
       lh.waitFor(spreds.replicatedStateIsReady(database, stateId, newServers));
 
-      lh.waitFor(lh.replicatedLogParticipantsFlag(database, stateId, {
+      lh.waitFor(lpreds.replicatedLogParticipantsFlag(database, stateId, {
         [newParticipant]: {allowedInQuorum: true, allowedAsLeader: true, forced: false},
       }));
     },
@@ -213,8 +214,8 @@ const replicatedStateSuite = function () {
           target.leader = newLeader;
           return target;
         });
-      lh.waitFor(lh.replicatedLogLeaderTargetIs(database, stateId, newLeader));
-      lh.waitFor(lh.replicatedLogLeaderPlanIs(database, stateId, newLeader));
+      lh.waitFor(lpreds.replicatedLogLeaderTargetIs(database, stateId, newLeader));
+      lh.waitFor(lpreds.replicatedLogLeaderPlanIs(database, stateId, newLeader));
     },
 
     testUnsetLeader: function () {
@@ -247,7 +248,7 @@ const replicatedStateSuite = function () {
         });
       lh.waitFor(spreds.replicatedStateVersionConverged(database, stateId, 2));
       const newServers = _.without(servers, serverToBeRemoved);
-      lh.waitFor(lh.replicatedLogParticipantsFlag(database, stateId, {
+      lh.waitFor(lpreds.replicatedLogParticipantsFlag(database, stateId, {
         [serverToBeRemoved]: null,
       }));
       lh.waitFor(spreds.replicatedStateIsReady(database, stateId, newServers));
